fix: generate thumbnail for multilayer tiff
When there is more than one layer in a tiff file, ImageMagick's convert
will generate multiple files as output, with a -0, -1 (and so on)
suffix. This causes a 404 error when specify goes to search for the
thumbnails, and produces excess images in the thumbnails directly. This
commit checks if there are multiple layers in the tiff, and if so, uses
the first layer to generate the thumbnail (similar to how PDF thumbnails
are handled).